refactor: CardFactory 피드백 반영
